Removed Leonardo-era workaround. Teensy is fast enough not to need it.
Normally, the shift register would update values on the rising edge,
while the SNES would sample on the falling edge. The Leonardo wasn't
fast enough to get data ready from a rising edge interrupt, so it
interrupted on the falling edge and tried to prepare the NEXT value
while the SNES is busy sampling. The Teensy is too fast, and actually
replaced the value the SNES observed... so now it's behaving more like
the proper shift register would.